+++ module-domain/src/test/java/com/depromeet/mock/image/FakeS3ImageManager.java
@@ -1,4 +1,4 @@
-package com.depromeet.mock;
+package com.depromeet.mock.image;
 
 import com.depromeet.image.port.out.s3.S3ManagePort;
