+++ module-domain/src/test/java/com/depromeet/mock/member/FakeMemberRepository.java
@@ -1,4 +1,4 @@
-package com.depromeet.mock;
+package com.depromeet.mock.member;
 
 import com.depromeet.member.domain.Member;
 import com.depromeet.member.domain.MemberGender;
